+++ core/base/src/jvmCommonMain/kotlin/nl/adaptivity/xmlutil/XmlStreaming.jvm.kt
@@ -129,9 +129,15 @@ internal actual object XmlStreaming : IXmlStreaming {
     actual override val platformDOMImplementation: PlatformDOMImplementation
         get() = DOMImplementationImpl.delegate
 
+    @Deprecated("Platform nodes are supertypes", level = DeprecationLevel.HIDDEN)
     @ExperimentalXmlUtilApi
     actual override fun newReader(source: Node): XmlReader {
-        return DomReader(source, false)
+        return DomReader(source, true)
+    }
+
+    @ExperimentalXmlUtilApi
+    actual override fun newReader(source: PlatformNode): XmlReader {
+        return DomReader(source)
     }
 
     fun newReader(inputStream: InputStream): XmlReader {