+++ Changelog.md
@@ -39,7 +39,8 @@ Changes:
   for strings already in memory.
 - Per the XML standard processing instructions are actually allowed in the XML
   body. Allow them here.
-- Add a newWriter implementation that targets platform nodes.
+- Add a newWriter implementation that targets platform nodes. Make newReader
+  consume platform nodes
 
 Fixes:
 - Make ElementSerializer work better in an existing document context by